ENH Refactor Element
