+++ client/src/components/ElementEditor/ElementContext.js
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+export const ElementContext = createContext(null);
